Added the `Help` command.

+++ Assistant.AI/Commands/Parsing/GuildPrefixResolver.cs
@@ -32,7 +32,7 @@ public class GuildPrefixResolver : IPrefixResolver {
             using var scope = ServiceManager.ServiceProvider!.CreateScope();
             SqliteDatabaseContext databaseContent = scope.ServiceProvider.GetRequiredService<SqliteDatabaseContext>();
 
-            GuildData? guildData = databaseContent.GuildDataSet.FirstOrDefault(guild => (ulong)guild.GuildId == message.Channel.Guild.Id);
+            GuildData? guildData = databaseContent.GuildDataSet.FirstOrDefault(guild => guild.GuildId == message.Channel.Guild.Id);
             guildData ??= new GuildData();
 
             if(string.IsNullOrWhiteSpace(guildData.Options.Prefix)) {
